+++ src/components/app-header/app-header.css
@@ -1,1 +1,5 @@
-
+.app-header{
+    display: flex;
+    justify-content: space-between;
+    align-items: baseline;
+}
