+++ src/components/app-header/app-header.js
@@ -1,8 +1,13 @@
 import React from "react";
 import "./app-header.css";
 
-const AppHeader = () => {
-  return <h1>My ToDo List</h1>;
+const AppHeader = ({ toDo, done }) => {
+  return (
+    <div className="app-header">
+      <h2>My ToDo List</h2>
+      <h5> {toDo} more to do, {done} done </h5>
+    </div>
+  );
 };
 
 export default AppHeader;
