+++ src/components/app/index.js
@@ -1,0 +1,2 @@
+import App from "./app";
+export default App;
