+++ src/components/item-status-filter/item-status-filter.js
@@ -1,14 +1,14 @@
-import React from 'react';
+import React, { Component } from "react";
 import "./item-status-filter.css";
-
-const ItemStatusFilter = ()=>{
+// create a class to get the internal state of the component
+export default class ItemStatusFilter extends Component {
+  render() {
     return (
-    <div className='btn-group  align-buttons'>
-        <button className='btn btn-info'>All</button>
-        <button className='btn btn-outline-secondary'>Active</button>
-        <button className='btn btn-outline-secondary'>Done</button>
-    </div>
-    )
-};
-
-export default ItemStatusFilter;
+      <div className="btn-group  align-buttons">
+        <button className="btn btn-info">All</button>
+        <button className="btn btn-outline-secondary">Active</button>
+        <button className="btn btn-outline-secondary">Done</button>
+      </div>
+    );
+  }
+}
